requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
dj_rest_auth
drf-yasg
django-filter
